requirements.txt
torch~=1.13.0a0+340c4120.nv22.6
torchvision~=0.13.0a0+da3794e
torchaudio
wheel
tqdm~=4.67.1
evo
scipy~=1.10.1
open3d~=0.16.0
gdown
tensorboard
opencv-python~=4.11.0.86
pyyaml~=6.0.2
numpy~=1.24.4
lietorch~=0.2
argparse~=1.4.0
setuptools~=44.0.0
moderngl~=5.12.0
matplotlib~=3.7.5
pillow~=10.4.0
requests~=2.32.4
pandas~=2.0.3
jinja2~=3.1.6